+++ lib/google_speech.dart
@@ -1,4 +1,7 @@
+export 'config/recognition_config.dart';
+export 'config/recognition_config_v1.dart';
+export 'config/recognition_config_v1p1beta1.dart';
+export 'config/streaming_recognition_config.dart';
 export 'exception.dart';
-export 'recognition_config.dart';
 export 'speech_client_authenticator.dart';
 export 'speech_to_text.dart';